renamed group webid input file as no longer a dialog
